requirements.txt
torch>=1.8.0
torchvision>=0.9.0
opencv-python>=4.5.0
numpy>=1.20.0
scikit-learn>=0.24.0
ensemble-boxes>=1.0.0
Pillow>=8.0.0
matplotlib>=3.0.0
ensemble_boxes>=1.0.0
# pytorch-grad-cam removed as it's not essential for core functionality
